refactor: transfer chapter awards page to app router
